More size_t fixes

+++ ACG/Scenegraph/DrawModes.cc
@@ -183,7 +183,7 @@ normalSource_(_normalSource)
 }
 
 
-DrawMode::DrawMode(unsigned int _index)
+DrawMode::DrawMode(size_t _index)
 {
   modeFlags_.reset();
   if ( _index >= modeFlags_.size() ) {
@@ -408,7 +408,7 @@ DrawMode DrawMode::operator~( ) const {
 
 
 
-int DrawMode::getIndex() const {
+size_t DrawMode::getIndex() const {
   if ( modeFlags_.count() == 1 ) {
     for ( size_t i = 0 ; i < modeFlags_.size() ; ++i ) 
      if ( modeFlags_[i] )
